implemented communication scheme between manager and client.
Changed given json interface to be valid.
Elements attribute is now an array of dictionaries.
Cross Validation sets got replaced by data sets, because
cross validation is not supposed to be done on client.
Added manager_info and client_info attributes that provide
important information for the manager or client.