Static pools
My furthest calculation stoppen after in the 10th opt-step with the message that it could not open more files. This exception was thrown when a pool tried to spawn new processes. The pools seemed to not clean up and finish their processes correctly. Maybe a pool.join() after pool.close() would also have helped stopping the garbage collector to garbage the pool-var before pool.close() finished alls processes. The advantage of this way is that we don't have to create a new pool in every train- oder predict-call. The disadvantage is that all models of the same type share the same pool. But that should not be a major problem as creating more processes does not necessarily speed calculations up due to limited capacities by the calculating computer. Also you can pass a bigger pool_size per parameter to create a bigger pool.